Extract visibility assertions
