+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/ClearFilesRuleTest.java
@@ -8,7 +8,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 
 public class ClearFilesRuleTest {
