+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/DialogActivityTest.java
@@ -10,7 +10,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.interaction.BaristaClickInteractions.clickOn;
 import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNegativeButton;
 import static com.schibsted.spain.barista.interaction.BaristaDialogInteractions.clickDialogNeutralButton;
