+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/ListsScrollTest.java
@@ -9,8 +9,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.interaction.BaristaListInteractions.scrollListToPosition;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertNotExist;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertNotExist;
 import static com.schibsted.spain.barista.sample.ListsActivity.FRUITS;
 import static com.schibsted.spain.barista.sample.ListsActivity.IntentBuilder;
