+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/MenuSupportActionBarClickTest.java
@@ -7,7 +7,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.interaction.BaristaMenuClickInteractions.clickMenu;
 
 @RunWith(AndroidJUnit4.class)
