+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/ViewPagerTest.java
@@ -8,7 +8,7 @@ import org.junit.runner.RunWith;
 
 import static com.schibsted.spain.barista.interaction.BaristaViewPagerInteractions.swipeViewPagerBack;
 import static com.schibsted.spain.barista.interaction.BaristaViewPagerInteractions.swipeViewPagerForward;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 
 @RunWith(AndroidJUnit4.class)
 public class ViewPagerTest {
