+++ sample/src/androidTest/java/com/schibsted/spain/barista/sample/WrappedViewClickTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static com.schibsted.spain.barista.assertion.BaristaAssertions.assertDisplayed;
+import static com.schibsted.spain.barista.assertion.BaristaVisibilityAssertions.assertDisplayed;
 import static com.schibsted.spain.barista.internal.viewaction.PerformClickAction.clickUsingPerformClick;
 
 public class WrappedViewClickTest {
